# server/src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "UniVote API Documentation",
        "description": "REST API specification for the UniVote ACSES UMaT E-Voting System. Administrative actions require bearer tokens.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Base Server"
        }
    ],
    "paths": {
...
        "/candidates": {
            "get": {
                "summary": "Fetch all aspirants",
                "responses": {
                    "200": {
                        "description": "List of active aspirants."
                    }
                }
            },
            "post": {
                "summary": "Register candidate",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "electionId": {
                                        "type": "string",
                                        "example": "elec-12345"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Mensah"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "President"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#2e7d32"
                                    },
                                    "picture": {
                                        "type": "string",
                                        "description": "Base64 image representation"
                                    },
                                    "department": {
                                        "type": "string",
                                        "example": "Computer Science & Engineering"
                                    }
                                },
                                "required": [
                                    "electionId",
                                    "name",
                                    "position"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidate registered successfully."
                    }
                }
            }
        },
        "/candidates/{id}": {
            "delete": {
                "summary": "Remove candidate from ballot",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aspirant entry cleared."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    }
}